+++ OzzMarkdown.WPF/ViewModels/MainViewModel.cs
@@ -1,3 +1,6 @@
+using OzzMarkdown.Core.Extensions;
+using OzzMarkdown.Core.Models;
+using OzzMarkdown.WPF.Models;
 using OzzMarkdown.WPF.Services;
 using OzzWpf.Core.Commands;
 using OzzWpf.Core.Dialogs;
@@ -10,21 +13,30 @@ namespace OzzMarkdown.WPF.ViewModels;
 public class MainViewModel : AbstractViewModel
 {
     private readonly IFileDialogService _fileDialogService;
+    private ReleaseSource _releaseSource = new ReleaseSource();
+
 
     public MainViewModel() : this(new Win32FileDialogService()) { }
 
     public MainViewModel(IFileDialogService fileDialogService)
     {
         _fileDialogService = fileDialogService ?? throw new ArgumentNullException(nameof(fileDialogService));
+
+        CheckForUpdatesCommand = new RelayCommand(async () => await CheckForUpdatesAsync());
         OpenFileCommand = new RelayCommand(async () => await OpenMarkdownFileAsync());
         ShowAboutCommand = new RelayCommand(ShowAboutDialog);
+
+        //_ = Task.Run(CheckForUpdatesAsync);
     }
+
+
+    public RelayCommand CheckForUpdatesCommand { get; }
 
     public RelayCommand OpenFileCommand { get; }
 
     public RelayCommand ShowAboutCommand { get; }
 
-
+    public GitHubRelease? LatestRelease { get; private set; }
 
     public string? CurrentFilePath
     {
@@ -54,6 +66,20 @@ public class MainViewModel : AbstractViewModel
     }
     private bool? _generateToc;
 
+    public bool HasNewerVersion
+    {
+        get => _hasNewerVersion;
+        set
+        {
+            if (_hasNewerVersion != value)
+            {
+                _hasNewerVersion = value;
+                RaisePropertyChanged(nameof(HasNewerVersion));
+            }
+        }
+    }
+    private bool _hasNewerVersion;
+
     public string? MarkdownContent
     {
         get => _markdownContent;
@@ -67,6 +93,25 @@ public class MainViewModel : AbstractViewModel
         }
     }
     private string? _markdownContent;
+
+    private async Task CheckForUpdatesAsync()
+    {
+        try
+        {
+            LatestRelease = await _releaseSource.GetGitHubReleaseAsync();
+            if (LatestRelease != null && !string.IsNullOrEmpty(LatestRelease.TagName))
+            {
+                // Compare the current version with the latest release version
+                Version currentVersion = new Version(_releaseSource.CurrentVersion);
+                Version latestVersion = new Version(LatestRelease.TagName.TrimStart('v'));
+                HasNewerVersion = latestVersion > currentVersion;
+            }
+        }
+        catch (Exception ex)
+        {
+            // Handle exceptions (e.g., log them or show a message to the user)
+        }
+    }
 
     private async Task OpenMarkdownFileAsync()
     {
@@ -88,9 +133,9 @@ public class MainViewModel : AbstractViewModel
         }
     }
 
-    private static void ShowAboutDialog()
+    private void ShowAboutDialog()
     {
-        var aboutDialog = new AboutDialog();
+        var aboutDialog = new AboutDialog(_releaseSource);
         if (Application.Current?.MainWindow != null)
         {
             aboutDialog.Owner = Application.Current.MainWindow;